refactor: made changes to the spaces section
