+++ src/App.tsx
@@ -2,7 +2,7 @@ import "./App.css";
 import Header from "./components/Header";
 import Hero from "./components/Hero";
 import About from "./components/About";
-import TalkEvent from "./components/TalkEvent";
+import Talks from "./components/Talks";
 import Footer from "./components/Footer";
 import FloatingBtn from "./components/floatingBtn";
 import FAQs from "./components/FAQs";
@@ -16,7 +16,7 @@ function App() {
       <FloatingBtn />
       <main>
         <About />
-        <TalkEvent />
+        <Talks />
         <FAQs />
       </main>
       <Footer />
