+++ src/components/Talks/index.tsx
@@ -1,0 +1,12 @@
+import Spaces from "./spaces";
+import Speaker from "./speakers";
+
+const Talks = () => {
+  return (
+    <section className="relative z-20 w-full bg-[#2F20BF]">
+      <Spaces />
+      <Speaker />
+    </section>
+  );
+};
+export default Talks;
