+++ tailwind.config.ts
@@ -24,22 +24,23 @@ export default {
       },
       textColor: {
         "primary-black": "#1F1F1F",
-        "white": "#FFF",
-        "dark": "#000000",
+        white: "#FFF",
+        dark: "#000000",
         gray: "#666",
-        "purple": "#4334D3",             
+        purple: "#4334D3",
       },
       backgroundColor: {
         "dark-purple": "#2F20BF",
-        "white": "#FFF",
-        "peach": "#ECC89D"
+        white: "#FFF",
+        peach: "#ECC89D",
       },
       backgroundImage: {
         "lg-banner": "url('/src/assets/lg-banner.png')",
-        "spaces": "url('/src/assets/Spaces.png')",
+        spaces: "url('/src/assets/Spaces.png')",
         "sm-banner": "url('/src/assets/sm-banner.png')",
         "lg-aboutSectionBg": "url('/src/assets/lg-about-section-bg.svg')",
         "sm-aboutSectionBg": "url('/src/assets/sm-about-section-bg.svg')",
+        "speaker-bg": "url('/src/assets/speak-image.png')",
       },
       borderRadius: {
         "6xl": "32px",
